+++ jukebox-app/src/main/res/layout/nowplaying_local.xml
@@ -21,7 +21,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content" >
 
-        <SurfaceView
+        <VideoView
             android:id="@+id/surfaceview"
             android:layout_width="fill_parent"
             android:layout_height="wrap_content"
